requirements.txt
# Core libraries for Q&A chatbot
streamlit==1.29.0
langchain==0.0.350
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langchain-community==0.0.11

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0
pymupdf==1.23.8

# Vector stores and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.18

# Text processing and utilities
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.4

# Additional utilities
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# For evaluation and metrics
rouge-score==0.1.2
bert-score==0.3.13
